requirements.txt
streamlit
fastai
pillow
numpy
requests
llama-cpp-python
chromadb
google-generativeai
sentence-transformers
langchain
langchain-google-genai
